Deepen config and topic store modules
