+++ packages/column-live-view-engine/src/raw-query-compiler.ts
@@ -454,7 +454,6 @@ const matchesFilter = (value: unknown, filter: unknown): boolean => {
     }
     return eq !== undefined || oneOf !== undefined || notEqual !== undefined;
   }
-  /* v8 ignore next -- runtime validation rejects scalar object filters before evaluation. */
   if (!isOperatorFilterObject(filter)) {
     return valuesEqual(value, filter);
   }
