+++ packages/config/src/runtime.ts
@@ -1,6 +1,8 @@
 import { Config } from "effect";
 import type { RuntimeEnvironmentConfig } from "./index";
 
+export type { RuntimeEnvironmentConfig } from "./index";
+
 export const runtimeEnvironmentConfig: RuntimeEnvironmentConfig = {
   websocketPort: Config.number("VIEW_SERVER_WEBSOCKET_PORT"),
   tcpPublishPort: Config.number("VIEW_SERVER_TCP_PUBLISH_PORT"),
